% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unify_coords.R
\name{unify_coords}
\alias{unify_coords}
\title{Unify coordinates in different format to decimal coordinates}
\usage{
unify_coords(x)
}
\arguments{
\item{x}{a coordinate as string}
}
\value{
a unified, numeric decimal coordinate
}
\description{
Unify coordinates in different format to decimal coordinates
}
