% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unify_coords.R
\name{unify_coords}
\alias{unify_coords}
\title{Unify coordinates in different format to decimal coordinates}
\usage{
unify_coords(x)
}
\arguments{
\item{x}{coordinates as a string vector}
}
\value{
unified, numeric decimal coordinates
}
\description{
This function can convert from various (and mixed) coordinate specifications
into the unified decimal degree format.
}
\details{
Examples for supported formats:

Degree-minute-second formats
\itemize{
\item 23°30'30''N
\item N 23°30'30"
\item 23 ° 30´ 30.521′′ N
\item N -23°30' 30.521"
}

Degree-decimal-minute formats
\itemize{
\item 23°12'
\item 23°12.00'
\item 23°12.00`
\item 23°12,00`
\item N 23°12,612`
\item -23°12.612´S
}

Decimal degrees
\itemize{
\item 23
\item 23°
\item 23.1
\item N23,1°
\item -23.12° N
}
}
\examples{
unify_coords(c("49°40'46.148``N", "E 48.232°"))
}
